feat: home page 교체 및 테스트 api
